move stat interrupt manager class
